separate app ui configs slice
